requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
mysqlclient==2.2.5
packaging==24.2
sqlparse==0.5.1
tzdata==2024.2